renamed sup-parts of a encapsed_string to string_part
When they're aliased as $.string you'll get string nodes in the tree with different origin and with different semantic meening.

"$foo'bar'$baz"  and 'bar' will both produce a $.string with the content 'bar', but you'll have to check if it is a direct child of encapsed_string to know if the single quotes are a part of the string or not.